\name{Nitrite}
\alias{Nitrite}
\title{Nitrite in Water}
\usage{
Nitrite (Conc_std_sol, E1, E1_s, E0, EB1)
}
\arguments{
\item{Conc_std_sol}{Concentration of standard solution}
\item{E1}{Absorbance of standard with reagent}
\item{E1_s}{Absorbance of sample with reagent}
\item{E0}{Absorbance of sample without NEED reagent}
\item{EB1}{Absorbance of distilled water and reagent}
}
\value{
Nitrite (micro gm/L)
}
\description{Nitrite is the key part of the nitrogen cycle in water. During denitrification, which occurs in low oxygen conditions, nitrite can build up. High levels of nitrite are harmful to fish, shrimp, and other aquatic organisms, making it important to manage for their health and well-being.


Principle of analysis:


Sulphanilamide is used as the amino compounds which is coupled with n-(1-napthyl)-ethylene diamine dihydrochloride. This reaction leads to the formation of azo dye formed is proportional to the initial concentration of nitrite. Absorbance is measured 543nm in spectrophotometer. APHA(2017,ISBN:9780875532875).
}
\examples{
 Nitrite (1000, 0.30, 0.20, 0.10, 0.05)
}
\references{
APHA, 2017. Standard methods for examination of water and wastewater, 23rd edn. APHA, AWWA, WPCF, Washington
}
\keyword{Nitrite in Water}
