\name{Salinity}
\alias{Salinity}
\title{Salinity in Water using Mohr-Knudsen Method}
\usage{
Salinity (chl_std_seawater, T1, T2)
}
\arguments{
\item{chl_std_seawater}{Chlorinity of standard seawater}
\item{T1}{Volume of silver nitrate required to titrate a sample}
\item{T2}{Volume of silver nitrate required to titrate standard seawater}
}
\value{
Salinity in Water(gm/L)
}
\description{Salinity refers to the concentration of dissolved salts in water, typically measured in parts per thousand (ppt). APHA(2017,ISBN:9780875532875).
}
\examples{
Salinity (19, 25, 30)
}
\references{
APHA, 2017. Standard methods for examination of water and wastewater, 23rd edn. APHA, AWWA, WPCF, Washington
}
\keyword{Salinity in Water}
