\name{ToW}
\alias{ToW}
\title{Transparency of Water using Secchi Disk}
\usage{
ToW (A, B)
}
\arguments{
\item{A}{Depth of disappearance}
\item{B}{Depth of reappearance}
}
\value{
Transparency of Water(cm)
}
\description{The transparency of water in aquatic systems refers to the degree to which light can penetrate and travel through the water without being absorbed or scattered. It is also referred to as water clarity. APHA(2017,ISBN:9780875532875).
}
\examples{
ToW (90, 30)
}
\references{
APHA, 2017. Standard methods for examination of water and wastewater, 23rd edn. APHA, AWWA, WPCF, Washington
}
\keyword{Transparency of Water}
